Updated to a more sensable file structure
